Implement an 'AtEndOfLine' method.
We need a way to determine when we're at the end of a line in the LineOfText.
Because the pointer stored in the LineOfText is used to point to the beginning
and end of the identifiers, we need to point to one past the end to get the
correct length.